autoservice-app/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.6
click==8.1.7
itsdangerous==2.1.2
blinker==1.7.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.2
psycopg2-binary
python-dotenv==1.0.0

gunicorn==22.0.0
cryptography==41.0.7